refactor(NotificationTypeLegend.js): Moved this file from the common location to the audit table package as it's the only place that needs this data
