+++ src/main/js/dynamic/loaded/audit/NotificationTypeLegend.js
